Move integer size and position 2D to math package
Move the integer size 2D and integer position 2D to the math package.
