+++ illud/math/integer_size_2d.py
