+++ mocks/terminal_cursor.py
@@ -1,7 +1,7 @@
 """A mock terminal cursor."""
 from unittest.mock import patch
 
-from illud.integer_position_2d import IntegerPosition2D
+from illud.math.integer_position_2d import IntegerPosition2D
 from illud.terminal_cursor import TerminalCursor
